todo-api build number based on number of git commits.
The docker image version tag for the todo-api is now manually input (through Jenkinsfile updates) for the Major.Minor versions and the build number is based on the number of git commits. This is usually a better system then either relying on someone to update the build number manually, or tying the build number to something in the CI/CD environment because it is instead exactly tied to a specific commit.