Added dummy AZURE_CLIENT_ID so avoid error during nginx start.
For now there is a AZURE_CLIENT_ID for the user-api brought up in the todo-ui which just defaults to a dummy value for a simple 'docker-compose up'. Ideally we should have all apps start up in test mode by default or we have a known, shared, Azure app which we use the client Id as the default for the AZURE_CLIENT_ID version.